requirements.txt
fastapi
uvicorn[standard]
jinja2
aiofiles
sqlalchemy
python-multipart
asyncpg
fpdf
openpyxl
aiogram>=3.0.0
pandas
greenlet
apscheduler
appium-python-client
alembic
python-dotenv